tui/window: refactor camera
