+++ src/simulation/mod.rs
@@ -1,5 +1,5 @@
 #[allow(clippy::module_inception)]
-pub mod simulation;
+mod simulation;
 mod control;
 
 pub use simulation::Simulation;
