+++ src/world/mod.rs
@@ -1,5 +1,5 @@
 #[allow(clippy::module_inception)]
-pub mod world;
+mod world;
 pub mod celestials;
 pub mod config;
 pub mod spaceship;
